requirements.txt
argcomplete==2.0.0
boto3==1.22.10
botocore==1.25.10
cachetools==4.2.4
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.3
dialogflow==2.0.0
durationpy==0.5
Flask==2.1.2
future==0.18.2
google-api-core==1.31.5
google-auth==1.35.0
google-cloud==0.34.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
hjson==3.0.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
packaging==21.3
pep517==0.12.0
pip-tools==6.6.0
placebo==0.8.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
troposphere==4.0.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wsgi-request-logger==0.4.6
zipp==3.8.0
